+++ changelog.md
@@ -3,6 +3,8 @@ Portuguese translation by Rodrigodd<br>
 DNA Sequencer upgrade<br>
 Gene Assembly cost increase to 200k<br>
 Synthesis now applies a bonus to auto crafting<br>
+Ambidextrous buffed<br>
+Ambidextrous now has a greater effect on auto crafting<br>
 
 #### 0.4.29 - 7/18/2019 ####
 Genetic modification system, customize your race with minor traits<br>